Add filter for preprocessing and classifier
We remove combined score, because combined score is the combination of
gestalt and feature score. Now we only use feature, CADD, Gestalt, Boqa
and Phenomize score.

Add filter in data class. We exclude the sample which has no gestalt
score. When we use function loadData(), the second parameter is the
field we want to filter. If the pathogenic gene of the sample doesn't
contain the value in this field, then we exclude this sample.

Add a filter in classifier. We can filter the feature which we don't
wnat to be trained by the classifier. In function classify(), the last
parameter is the list of features which will be ignored by classifier