Add script for drawing nan distribution
Add draw_nan_dist.py to check the nan distribution on all feature vector.
the output files will be in output/nan_dist/'name you type in argument'
nan_dist.png and pa_nan_dist.png are the figures which show the
distribution of nan frequency.
The prefix 'pa' means that it is for pathogenic gene.
nan_feature.png and pa_nan_feature.png are the figures which show the
number of sample which don't have the corresponding feature.
feature_count.png and pa_feature_count.png are the figures which show
the number of sample which have the corresponding feature
Each *.csv file has a list of sample and gene which have the
corresponding feature.