Add support for running on server
Add input argument -s to mention if we are running on server.
If we use -s, then we are running on server. The testing json file is
reading from stdin, and the output pedia score is to stdout.
Use logger instead of print function. The log will write to run.log.