Rename performance output file
Use rank.csv instead of rank_gene_1KG.csv. In rank.csv, we show the
rank of correct diagnosis of gene in the case.
Use count.csv instead of rank_1KG.csv. In count.csv, we show the count
of cases in the specific rank.
We use dataframe to store the pedia results instead of several arrays.